Move bill services
