+++ src/Services/Bill/BillService.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services;
+namespace App\Services\Bill;
 
 use App\Services\Cdr\CdrOutputInterface;
 use App\Services\Soap\ExceptionCreator;
